View delete post

See merge request lbaw/lbaw2021/lbaw2123!7

+++ app/Http/Controllers/PostController.php
@@ -5,8 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Models\Photo;
 use App\Models\Tag;
+use App\Models\AuthenticatedUser;
+use App\Models\Comment;
+use App\Policies\PostPolicy;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\DB;
+use Auth;
 
 class PostController extends Controller
 {
@@ -119,11 +124,52 @@ class PostController extends Controller
      */
     public function show($id)
     {
-        //$post = Post::find($id); -- ver se o post existe 
+
+        //Get currnt route
+        $route = \Route::current();
+
+        //If route {id} isnt int or post doesnt exist, redirect to notfound.
+        if(!is_numeric($route->parameter('id')))
+            return view('pages.pagenotfound',['user' => 'visitor','needsFilter' => 0]);
+        $post = Post::find($id);
+        if(!$post )
+            return view('pages.pagenotfound',['user' => 'visitor','needsFilter' => 0]);
+
+        //Verify if user is authenticated and if user is owner of post    
+        if(Auth::check())
+            $user = Auth::user()->id == $post->user_id? 'authenticated_owner' : 'authenticated_user';
+        else
+            $user = 'visitor';
+        $user = 'authenticated_user';
         
-        //checkar se está autenticado e se é o dono
-        //se o post existir vai buscar o necessario para mostrar o post e chama a sua view
-        return view('pages.post', ['user' => 'visitor', 'needsFilter' => 0] ); //['post'=> $post]
+        $post->timestamps = false;
+        $post->increment('n_views');
+
+        //Get owner
+        $USER = AuthenticatedUser::find($post->user_id);
+
+        //Get comment count,likes and dislikes
+        $comments = Comment::where('post_id',$id)->get()->count();
+        $votes = DB::table("vote_post")->where("post_id",$id);
+        $temp = $votes->get()->count();
+        $likes = $votes->where("like",true)->get()->count();
+        $dislikes = $temp - $likes;
+
+        //Get tags associated with current post
+        $tags = DB::select(DB::raw("select t.name
+        FROM post_tag,tag as t
+        WHERE post_tag.post_id=$id AND t.id = post_tag.tag_id;"));
+
+        //Get date and thumbnail path
+        $date = date("F j, Y", strtotime($post['created_at']));
+        $thumbnail = "/images/".$post->thumbnail;
+
+        //Generate metadata to send to view
+        $metadata = ['comments'=>$comments,'author'=>$USER['name'],'views' => $post->n_views,
+                     'likes' => $likes,'tags' => $tags,'date'=>$date,'thumbnail' => $thumbnail];
+
+        
+        return view('pages.post', ['user' => $user, 'needsFilter' => 0,'post' => $post,"metadata"=> $metadata] );
     }
 
     /**
@@ -204,15 +250,18 @@ class PostController extends Controller
     public function destroy($post_id)
     {
         //checkar se está autenticado
-
-        $post = Post::find($post_id);
-        if($post != null){
-            if ($post->delete()) {
-                return; //dar return da view da homepage
-            } else {
-                return; // dar return da view do post
+        if(Auth::check()){
+            $post = Post::find($post_id);
+            $this->authorize("delete",[Auth::user(),$post]);
+            if($post != null){
+                if ($post->delete()) {
+                    return ''; //dar return da view da homepage
+                } else {
+                    return 'post/' + $post_id; // dar return da view do post
+                }
             }
         }
+        return '';
     }
 
     /**
@@ -249,5 +298,41 @@ class PostController extends Controller
         else return; //nao sei para onde vai 
 
         
+    }
+
+    public function addSave($id){
+        $route = \Route::current();
+
+        //If route {id} isnt int or post doesnt exist, redirect to notfound.
+        if(!is_numeric($route->parameter('id')))
+            return '';
+        if(Auth::check()){
+            $post = Post::find($post_id);
+            $this->authorize("save",[Auth::user(),$post]);
+            if($post != null){
+                DB::table("saves")->insert([
+                    'user_id' => 1,
+                    'post_id' => $post_id
+                ]);
+                return 'SUCCESS';
+            }
+        }
+        return '';
+    }
+
+    public function deleteSave($id){
+        $route = \Route::current();
+        //If route {id} isnt int or post doesnt exist, redirect to notfound.
+        if(!is_numeric($route->parameter('id')))
+            return '';
+        if(Auth::check()){
+            $post = Post::find($post_id);
+            $save = DB::table("save")->where("post_id",$post_id)->where("user_id",Auth::user()->id);
+            if($post != null && $save != null){
+                if($save->delete())
+                    return 'SUCCESS';
+            }
+        }
+        return '';
     }
 }
